Fix download on script completion on Chrome
